requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.5.0
astunparse==1.6.3
attrs==24.2.0
AudioConverter==0.4.1
autopep8==1.6.0
Babel==2.10.1
boto3==1.34.18
botocore==1.34.18
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
cloudconvert==2.1.0
colorama==0.4.4
contourpy==1.3.0
cv2module==0.1.2
cvlib==0.2.7
cycler==0.11.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-storages==1.14.2
docopt==0.6.2
docx==0.2.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpeg==1.4
filelock==3.6.0
fire==0.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.33.3
gast==0.4.0
google-api-core==2.11.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-speech==2.20.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
googletrans==3.1.0a0
grpcio==1.56.0
grpcio-status==1.56.0
gTTS==2.5.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
ibm-cloud-sdk-core==3.15.1
ibm-watson==6.0.0
idna==2.10
imageio==2.35.1
importlib-metadata==6.5.0
imutils==0.5.4
itsdangerous==2.1.0
jax==0.4.33
jaxlib==0.4.33
Jinja2==3.0.3
jmespath==1.0.1
keras==3.5.0
kiwisolver==1.4.2
libclang==16.0.0
libretranslatepy==2.1.1
lxml==4.8.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.0
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
mysql-connector-python==8.3.0
namex==0.0.8
numpy==1.24.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
packaging==21.3
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.4.0
pipreqs==0.4.11
platformdirs==2.5.1
playsound==1.3.0
progressbar==2.5
proto-plus==1.22.3
protobuf==4.25.4
psycopg2==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyaudioconvert==0.0.5
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.3.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.0.8
pypdf==4.2.0
PyPDF2==3.0.1
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2022.1
pywin32==306
PyYAML==6.0.1
reportlab==4.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.8.1
rsa==4.9
s3transfer==0.10.0
scipy==1.14.1
screen_brightness_control==0.23.0
six==1.16.0
sniffio==1.2.0
sounddevice==0.5.0
sox==1.4.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
starlette==0.36.3
sympy==1.11.1
tensorboard==2.17.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.17.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tesseract==0.1.3
tkcalendar==1.6.1
toml==0.10.2
translate==3.6.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.9
uvicorn==0.28.1
virtualenv==20.14.1
vosk==0.3.32
watchfiles==0.21.0
websocket-client==1.1.0
websockets==12.0
Werkzeug==2.0.3
whitenoise==6.0.0
WMI==1.5.1
wrapt==1.14.1
yarg==0.1.9
zipp==3.15.0
